Release 2 | Beta 0 | Development 0 | Experimental 4
* Changed MeshLoader to output the error text from Assimp if
  a file fails to load
* Fixed bug: Crash when loading a tilemap with Tilemap and
             TilemapLayer
* Fixed bug: Crash when updating Text with an empty string
* Fixed bug: Explicit instantiations of VBO should have been
             after the defininitions in VBO.cpp

+++ UnnamedEngine/Source/core/gui/Font.cpp
@@ -60,7 +60,7 @@ void Font::setup(std::string path, unsigned int size, TextureParameters paramete
 		//Add onto the width
 		width += glyphSlot->bitmap.width + GLYPH_SPACING;
 		//Update the height to make it the same as the tallest letter
-		height = utils_maths::max(height, glyphSlot->bitmap.rows);
+		height = utils_maths::max(height, (int) glyphSlot->bitmap.rows);
 	}
 
 	if (! BaseEngine::usingVulkan()) {
